requirements.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.6.1
django-eav2==1.4.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-location-field==2.7.1
django-mptt==0.14.0
Pillow==10.0.0
sqlparse==0.4.4
Faker~=19.2.0
requests~=2.31.0